chore: Move DeclarativeToaster to atoms folder
