+++ src/domain/gift-cards/manage/index.tsx
@@ -4,7 +4,7 @@ import React, { useEffect, useState } from "react"
 import { FieldValues } from "react-hook-form"
 import toast from "react-hot-toast"
 import Spinner from "../../../components/atoms/spinner"
-import Toaster from "../../../components/declarative-toaster"
+import Toaster from "../../../components/atoms/declarative-toaster"
 import FormToasterContainer from "../../../components/molecules/form-toaster"
 import useNotification from "../../../hooks/use-notification"
 import Medusa from "../../../services/api"
