+++ src/domain/products/edit.tsx
@@ -3,7 +3,7 @@ import React, { useEffect, useState } from "react"
 import { FieldValues } from "react-hook-form"
 import toast from "react-hot-toast"
 import Spinner from "../../components/atoms/spinner"
-import Toaster from "../../components/declarative-toaster"
+import Toaster from "../../components/atoms/declarative-toaster"
 import FormToasterContainer from "../../components/molecules/form-toaster"
 import { FeatureFlagContext } from "../../context/feature-flag"
 import useNotification from "../../hooks/use-notification"
